requirements.txt
Django
gunicorn
psycopg2-binary
python-dateutil